Task #7335. Changed the super class of the _InvoiceLine to InvoicePayableLine; changed types of some money properties to Money; some refactoring
git-svn-id: https://svn.ish.com.au/ish/willow/trunk@11497 fb2e276d-fae9-0310-b578-b71ac85840b9

+++ frameworks/common-model/src/main/java/ish/oncourse/model/Discount.java
@@ -248,7 +248,7 @@ public class Discount extends _Discount {
 					if (LOG.isDebugEnabled()) {
 						LOG.debug("applying discount % : " + getDiscountRate());
 					}
-					result = courseClass.getFeeExGst().multiply(
+					result = courseClass.getFeeExGst().toBigDecimal().multiply(
 							getDiscountRate());
 
 					if (getMaximumDiscount() != null
